+++ CHANGELOG.rst
@@ -4,8 +4,7 @@ CHANGELOG
 5.1.24 (unreleased)
 -------------------
 
-- Provide way to restrict where objects can be created
-  (allowed_parent_types) [lferran]
+- Provide way to define a content type not globally addable [lferran]
 
 5.1.23 (2019-12-11)
 -------------------